requirements.txt
spotipy
python-dotenv
requests
deezer-python
rapidfuzz
pandas
